fix concurrency bug:
Previously, some writes to the db were not completed/flushed by the time a
worker was fetching a job off the queue.  The db.get would return no entry for
the job because the prior write hadn't finnished yet.  So all jobs in the
queue up to the first job that was written to the db were being skipped/wiped.
But there was also a queue = nil assignment that occured if the first "ready"
job was nil (added due to a very wrong assumption) that would result in the
entire queue being cleared if the first non-queued jobid returned nil from the
database.  This fixes that by looping over the queue skipping nil or
not-queued jobs and splicing out the first ready job only.  The queue is never
reset to nil regardless of whether or not any jobs are found to be ready for
the worker to start running.