change queue to hold Job* instead of ids...
...for reduced db querying and faster overall performance and correctness.
There were some strange race conditions before that resulted in the queue
having many jobs (job ids) in it but the server would be stuck telling all
workers that there was no work in the queue.  The only way this could have
bene happening as far as I can tell is that all jobs in the queue were either
not retrievable from the database validly or they were not marked with Status
== StatusQueued despite being in the queue.  I'm not sure how this condition
has been arising, but hopefully, this might fix it.