fix db syncing bugs:
Previously, the database would not have some jobs synced/flushed by the time a
worker would try to open up the network conn/stream for the output data files.
Workers open this stream immediately as soon as they *start* a job to stream
output file data directly.  So sometimes the server db wouldn't have flushed
data by this time.  The output file dumping, due to sloppy implementation,
was previously doing a full job retrieval which would fail.  The failure would
propogate back to the worker and the worker would immediately return the job
as done but without any output file data and would then move on to the next
job without even terminating the job that it failed that was still running.

This update fixes this by:

* create a map holding pointers in memory of running jobs.
* retrievals, beat checks, beat timeouts consult the new running job map
* job outfile push streaming no longer performs a job retrieval.
* worker now waits for job to finish even if job outfile stream connection fails.