fix missed db sync problem from previous updates:
Previously, the job submitter would also try to retrieve the outfiles and
their data after it was reported to them that their job was complete.  The
original submitter would then query the rest http api before the db had synced
sometimes - and the outfiles api endpoint for GET requests still errored if
jobs weren't found in the db.  The POST portion of the endpoint was fixed in
the prev commit, but the GET was missed and is fixed here.